Created CSS and HTML doc
Fully created all container blocks according to the design. The approach was
to create 4 containers and customize each container according to the design
template. To create a margin left and right, a div in a div was created with
the inner div having size of 80%. Not optimal but this was the quickest solution I came up with.